+++ Matrix/MatrixType.h
@@ -7,24 +7,24 @@
 class MatrixType
 {
 protected:
-    int64_t fRows;
-    int64_t fCols;
+    int fRows;
+    int fCols;
     
 public:
-    MatrixType(int64_t rows, int64_t cols);
+    MatrixType(int rows, int cols);
     virtual ~MatrixType();
 
-    virtual void AddValueMatrix(int64_t &row, int64_t &col, double &val) = 0;
-    virtual void PutValueMatrix(int64_t &row, int64_t &col, double &val) = 0;
-    virtual double GetValueMatrix(int64_t &row, int64_t &col) = 0;
+    virtual void AddValueMatrix(int &row, int &col, double &val) = 0;
+    virtual void PutValueMatrix(int &row, int &col, double &val) = 0;
+    virtual double GetValueMatrix(int &row, int &col) = 0;
 
-    virtual void AddValueRhs(int64_t &row, double &val) = 0;
-    virtual void PutValueRhs(int64_t &row, double &val) = 0;
-    virtual double GetValueRhs(int64_t &row) = 0;
+    virtual void AddValueRhs(int &row, double &val) = 0;
+    virtual void PutValueRhs(int &row, double &val) = 0;
+    virtual double GetValueRhs(int &row) = 0;
 
-    virtual void AddValueSolution(int64_t &row, double &val) = 0;
-    virtual void PutValueSolution(int64_t &row, double &val) = 0;
-    virtual double GetValueSolution(int64_t &row) = 0;
+    virtual void AddValueSolution(int &row, double &val) = 0;
+    virtual void PutValueSolution(int &row, double &val) = 0;
+    virtual double GetValueSolution(int &row) = 0;
 
     virtual void PrintMatrix() = 0;
     virtual void PrintRhs() = 0;
